refactor(web): 페이지 조립 구조 정리, 기존_pages에 있던 검색 전용 ui,MOCK데이터를 features/space-search/로 이동
