+++ apps/web/src/_pages/space-search/index.ts
@@ -1,1 +1,2 @@
-export { SpaceSearchPage } from "./ui/space-search-page";
+export { SpaceSearchHeader } from "./ui/space-search-header";
+export { SpaceSearchHero } from "./ui/space-search-hero";
