+++ apps/web/src/features/space-search/index.ts
@@ -1,0 +1,3 @@
+export { parseSpaceSearchQueryState } from "./model/search-params";
+export type { SpaceSearchPagination, SpaceSearchQueryState } from "./model/types";
+export { SpaceSearchResultsSection, SpaceSearchToolbarSection } from "./ui/space-search-sections";
