+++ apps/web/src/features/space-search/model/mocks.ts
@@ -1,0 +1,101 @@
+import type { SpaceCardItem } from "./types";
+
+export const SPACE_SEARCH_MOCK_ITEMS: SpaceCardItem[] = [
+  {
+    category: "운동/건강",
+    currentParticipants: 4,
+    deadlineLabel: "오늘 21시 마감",
+    district: "강남구",
+    id: "space-card-1",
+    imageAlt: "밝고 시티한 공간",
+    imageSrc: "https://www.figma.com/api/mcp/asset/fa6ac266-bd89-4abd-ab65-b3c8606c015e",
+    maxParticipants: 20,
+    metaChips: [
+      { id: "date", label: "1월 7일" },
+      { id: "time", label: "17:30" },
+    ],
+    title: "저녁 스피닝 스트레칭",
+  },
+  {
+    category: "취미/여가",
+    currentParticipants: 19,
+    deadlineLabel: "오늘 21시 마감",
+    district: "중구",
+    id: "space-card-2",
+    imageAlt: "따뜻한 무드의 실내 공간",
+    imageSrc: "https://www.figma.com/api/mcp/asset/4625a75b-412c-4bb7-b1c0-a24215fd36f6",
+    isLiked: true,
+    maxParticipants: 20,
+    metaChips: [
+      { id: "date", label: "1월 7일" },
+      { id: "time", label: "17:30" },
+    ],
+    status: {
+      label: "개설 확정",
+    },
+    title: "직장 독서 모임 만들기",
+  },
+  {
+    category: "취미/여가",
+    currentParticipants: 12,
+    deadlineLabel: "오늘 21시 마감",
+    district: "용산구",
+    id: "space-card-3",
+    imageAlt: "감각적인 카페 내부",
+    imageSrc: "https://www.figma.com/api/mcp/asset/a4b7bd93-bd2b-493b-ab38-2c392ac3dc00",
+    maxParticipants: 20,
+    metaChips: [
+      { id: "date", label: "1월 7일" },
+      { id: "time", label: "17:30" },
+    ],
+    status: {
+      label: "개설 확정",
+    },
+    title: "카페 투어 멤버 모집",
+  },
+  {
+    category: "스터디",
+    currentParticipants: 4,
+    deadlineLabel: "2일 후 마감",
+    district: "동작구",
+    id: "space-card-4",
+    imageAlt: "창이 넓은 회화 공간",
+    imageSrc: "https://www.figma.com/api/mcp/asset/fe2081f5-580a-4342-9e3f-1b0b1c39d75d",
+    maxParticipants: 20,
+    metaChips: [
+      { id: "date", label: "1월 7일" },
+      { id: "time", label: "17:30" },
+    ],
+    title: "영어 로스터리 회화방",
+  },
+  {
+    category: "운동/건강",
+    currentParticipants: 12,
+    deadlineLabel: "오늘 21시 마감",
+    district: "중구",
+    id: "space-card-5",
+    imageAlt: "단체 스트레칭 공간",
+    imageSrc: "https://www.figma.com/api/mcp/asset/a14fbce8-49c1-452b-ae78-f58297c88024",
+    maxParticipants: 20,
+    metaChips: [
+      { id: "date", label: "1월 7일" },
+      { id: "time", label: "17:30" },
+    ],
+    title: "국민 체조 함께 하기",
+  },
+  {
+    category: "기타",
+    currentParticipants: 12,
+    deadlineLabel: "오늘 21시 마감",
+    district: "마포구",
+    id: "space-card-6",
+    imageAlt: "음악 크루 테마 이미지",
+    imageSrc: "https://www.figma.com/api/mcp/asset/ce86979a-9059-4f6a-b62f-9403acb6c0e5",
+    maxParticipants: 20,
+    metaChips: [
+      { id: "date", label: "1월 7일" },
+      { id: "time", label: "17:30" },
+    ],
+    title: "음악 리스닝 클럽",
+  },
+];
